Compile netcdf manually to modify netcdf.h
Glen Hansen mentioned that we may be experiencing a problem
with Albany for large files because some header options in
netcdf were not explicitly set to match the Albany settings.

This update compiles netcdf manually with those settings. But
we have yet to test whether or not we see improvement.